+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ### Breaking Changes
 
-- [#274](https://github.com/okta/okta-react/pull/274)
+- [#275](https://github.com/okta/okta-react/pull/275)
   - `SecureRoute` should be imported from `@okta/okta-react/react-router-5`
 
 ### Features
 
-- [#274](https://github.com/okta/okta-react/pull/274)
+- [#275](https://github.com/okta/okta-react/pull/275)
   - Adds new component `<AuthRequired>` for integration with `react-router 6.x`. It should be imported from `@okta/okta-react/react-router-6`
 
 # 6.8.0